monitoring-api/requirements.txt
flask
pyodbc
python-dotenv
requests
gunicorn